Updated the dishwasher interface between driver and manager
The driver is now responsible to create the energy profile that is
corresponding to the program. Also the control parameter now support
explicit starting of a specific program at a specified time.

This also means that the driver is responsible to start the dishwasher at
the selected time, if the hardware doesn't support that option.