Fixed a possible deadlock situation.
The wrapper is calling isConnected on the MatchingPort, which locks. If at
the same time an other thread tries to send a message, for which you need
to lock the wrapper to send the notification, you could end up in a
deadlock (until now seen once in the testcase). This should fix that.