The ConnectionManager now uses the scheduling from the FlexiblePowerContext
Before this, each Endpoint had its own thread to handle messages, but this resulted in a lot of threads being created.
To limit this, it now reuses the scheduling thread from the FlexiblePowerContext. This creates a thread for each bundle, which is already shared with all the other scheduling stuff. This should make the system faster on small computers, such as the Raspberry Pi, which are bad at handling a lot of threads.